+++ Mods ChoGGi/Add 5 Times The Resource To Deposits/metadata.lua
@@ -1,10 +1,12 @@
 return PlaceObj("ModDef", {
-	"title", "Add 5 Times The Resource To Deposits v0.2",
-	"version", 2,
-	"saved", 1534680000,
+	"title", "Add 5 Times The Resource To Deposits v0.3",
+	"version", 3,
+	"saved", 1539777600,
 	"id", "ChoGGi_Add5TimesTheResourceToDeposits",
 	"author", "ChoGGi",
-	"code", {"Script.lua"},
+	"code", {
+		"Code/Script.lua"
+	},
 	"image","Preview.png",
 	"steam_id", "1427609324",
 	"lua_revision", LuaRevision,