+++ Mods ChoGGi/Auto Empty Waste Storage/metadata.lua
@@ -1,14 +1,14 @@
 return PlaceObj("ModDef", {
 	"title", "Auto Empty Waste Storage v0.3",
 	"version", 3,
-	"saved", 1536148800,
-	"tags", "Misc",
+	"saved", 1539777600,
 	"id", "ChoGGi_AutoEmptyWasteStorage",
 	"author", "ChoGGi",
 	"image","Preview.png",
-	"code", {"Script.lua"},
+	"code", {
+		"Code/Script.lua"
+	},
 	"lua_revision", LuaRevision,
-
 	"steam_id", "1485526508",
 	"description", [[Automatically empties waste storage sites.