+++ Mods ChoGGi/Sample Sidepanel Controls/Script.lua
@@ -7,6 +7,7 @@ local PlaceObj = PlaceObj
 local function RemoveXTemplateSections(list,name)
 	local idx = table.find(list, name, true)
 	if idx then
+		list[idx]:delete()
 		table.remove(list,idx)
 	end
 end